requirements.txt
fastapi
uvicorn[standard]
python-dotenv
httpx
pdfplumber
numpy
scikit-learn
google-generativeai